thn debug

+++ macro/ME_PR_thn.cpp
@@ -49,7 +49,7 @@ funcWithJson(void, ME_PR_thn)(TString path_config = "../config.json") {
                   },
                   {"fEta", "fEta1"})
           .Define("fVtxZ_extended",
-                  [](const double &vtxZ, const RVec<float> &delta_phi) {
+                  [](const float &vtxZ, const RVec<float> &delta_phi) {
                     ROOT::RVec<double> vtxZ_extended;
                     for (size_t i = 0; i < delta_phi.size(); ++i) {
                       vtxZ_extended.emplace_back(vtxZ);
